extract ifOperatorCalcResult
