+++ src/App.js
@@ -3,6 +3,7 @@ import "./App.css";
 import ResultComponent from "./Components/ResultComponent";
 import KeyPadComponent from "./Components/KeyPadComponent";
 import { evaluate, round } from "mathjs";
+import ifOperatorCalcResult from "./ifOperatorCalcResult";
 
 class App extends React.Component {
   state = {
@@ -60,116 +61,13 @@ class App extends React.Component {
         button.key !== "AC" &&
         button.key !== ")"
       ) {
-        let keyPressed = button.key;
-        let result = this.state.result;
-
-        if ("(".includes(keyPressed)) {
-          if (result.length > 1) {
-            if ("0123456789".includes(result[result.length - 1])) {
-              result = result + keyPressed.replace("(", "*(");
-            }
-          }
-        } else if (result === "") {
-          if ("+-(0123456789".includes(keyPressed)) {
-            result = keyPressed;
-          }
-          if ("/*".includes(keyPressed)) {
-            result = "0" + keyPressed;
-          }
-          if (".".includes(keyPressed)) {
-            result = "0.";
-          }
-        } else if ("0123456789".includes(result[result.length - 1])) {
-          const numberOfOpenP = (result.match(/\(/g) || []).length;
-          const numberOfCloseP = (result.match(/\)/g) || []).length;
-          if (
-            "*/-+0123456789.".includes(keyPressed) ||
-            (keyPressed.key === ")" && numberOfOpenP > numberOfCloseP)
-          ) {
-            result = result + keyPressed;
-          }
-        } else if (result[result.length - 1] === "(") {
-          if ("+-(0123456789".includes(keyPressed)) {
-            result = result + keyPressed;
-          }
-          if (".".includes(keyPressed)) {
-            result = result + "0.";
-          }
-        } else if (result[result.length - 1] === ")") {
-          const numberOfOpenP = (result.match(/\(/g) || []).length;
-          const numberOfCloseP = (result.match(/\)/g) || []).length;
-          if ("/*-+".includes(keyPressed)) {
-            result = result + keyPressed;
-          }
-          if (")".includes(keyPressed) && numberOfOpenP > numberOfCloseP) {
-            result = result + keyPressed;
-          }
-        } else if ("0123456789)".includes(result[result.length - 2])) {
-          if ("/*+-".includes(result[result.length - 1])) {
-            if ("(0123456789".includes(keyPressed)) {
-              result = result + keyPressed;
-            }
-            if (".".includes(keyPressed)) {
-              result = result + "0.";
-            }
-            if ("+".includes(keyPressed) && result[result.length - 1] === "-") {
-              result =
-                result.substring(0, result.length - 1) +
-                "+" +
-                result.substring(result.length - 1 + 1);
-            }
-            if ("-".includes(keyPressed) && result[result.length - 1] === "+") {
-              result =
-                result.substring(0, result.length - 1) +
-                "-" +
-                result.substring(result.length - 1 + 1);
-            }
-            if ("/".includes(keyPressed) && result[result.length - 1] === "-") {
-              result =
-                result.substring(0, result.length - 1) +
-                "/" +
-                result.substring(result.length - 1 + 1);
-            }
-            if ("/".includes(keyPressed) && result[result.length - 1] === "+") {
-              result =
-                result.substring(0, result.length - 1) +
-                "/" +
-                result.substring(result.length - 1 + 1);
-            }
-            if ("*".includes(keyPressed) && result[result.length - 1] === "-") {
-              result =
-                result.substring(0, result.length - 1) +
-                "*" +
-                result.substring(result.length - 1 + 1);
-            }
-            if ("*".includes(keyPressed) && result[result.length - 1] === "+") {
-              result =
-                result.substring(0, result.length - 1) +
-                "*" +
-                result.substring(result.length - 1 + 1);
-            }
-          }
-        } else if ("/*+-(".includes(result[result.length - 2])) {
-          if ("-+".includes(result[result.length - 1])) {
-            if ("0123456789(".includes(keyPressed)) {
-              result = result + keyPressed;
-            }
-            if (".".includes(keyPressed)) {
-              result = result + "0.";
-            }
-          }
-        } else if ((result[result.length - 1] || "").includes(".")) {
-          if ("0123456789".includes(keyPressed)) {
-            result = result + keyPressed;
-          }
-        } else if ((result || "").includes(".")) {
-          if ("+-*/0123456789".includes(keyPressed)) {
-            result = result + keyPressed;
-          }
-        }
+        const [buttonPressed, result] = ifOperatorCalcResult(
+          button.key,
+          this.state.result
+        );
 
         this.setState({
-          operate: button.key,
+          operate: buttonPressed,
           result: result,
         });
       }
